refactor: move buttons to toolbar
